moved components out of folders
