+++ client/src/pages/Playground.jsx
@@ -14,10 +14,10 @@ import { Search, Users } from "lucide-react";
 import { useState } from "react";
 import toast, { Toaster } from "react-hot-toast";
 import { useMutation, useQuery, useQueryClient } from "react-query";
-import CategoryDetailsModal from "../components/categories/CategoryDetailsModal";
-import CategoryEditableModal from "../components/categories/CategoryEditableModal";
-import ProductDetailsModal from "../components/products/ProductDetailsModal";
-import ProductEditableModal from "../components/products/ProductEditableModal";
+import CategoryDetailsModal from "../components/CategoryDetailsModal";
+import CategoryEditableModal from "../components/CategoryEditableModal";
+import ProductDetailsModal from "../components/ProductDetailsModal";
+import ProductEditableModal from "../components/ProductEditableModal";
 import CategoriesAPI from "../services/CategoriesAPI";
 import ProductsAPI from "../services/ProductsAPI";
 import UsersAPI from "../services/UsersAPI";
